+++ braincube_connector/memory_base/nested_resources/mb_child.py
@@ -6,7 +6,7 @@ from braincube_connector.bases import base_entity
 class MbChild(base_entity.BaseEntity):
     """MbChild object provides an interface between a child and its memory bases."""
 
-    def __init__(self, bcid, name, metadata, path, memory_base):
+    def __init__(self, bcid, name, metadata, path, **kwargs):
         """Initialize a memory bases child.
 
         Args:
@@ -14,11 +14,11 @@ class MbChild(base_entity.BaseEntity):
             name: Usual name of the entity description.
             metadata: Raw metadata associated to the entity description.
             path: Path of the entity description on the server.
-            memory_base: Instance of the parent memory base.
+            **kwargs: Optional braincube_name or parent_entity.
         """
-        self.initialize(bcid, name, metadata, path, memory_base)
+        self.initialize(bcid, name, metadata, path, **kwargs)
 
-    def initialize(self, bcid, name, metadata, path, memory_base):
+    def initialize(self, bcid, name, metadata, path, **kwargs):
         """Initialize a memory bases child.
 
         Args:
@@ -26,10 +26,10 @@ class MbChild(base_entity.BaseEntity):
             name: Usual name of the entity description.
             metadata: Raw metadata associated to the entity description.
             path: Path of the entity description on the server.
-            memory_base: Instance of the parent memory base.
+            **kwargs: Optional braincube_name or parent_entity.
         """
-        super().initialize(bcid, name, metadata, path)
-        self._memory_base = memory_base
+        super().initialize(bcid, name, metadata, path, **kwargs)
+        self._memory_base = self._parent_entity
 
     def get_memory_base(self) -> "MemoryBase":  # type: ignore  # noqa
         """Get the entity's parent memory bases.